+++ src/db-v1/archive.ts
@@ -1,4 +1,4 @@
-import { model, t } from '../db';
+import { model, t } from '../db-v1';
 
 export const Archive = model('wal_archives', {
   id: t.pk(),
