+++ src/db-v1/certs.ts
@@ -1,4 +1,4 @@
-import { model, t } from '../db';
+import { model, t } from '../db-v1';
 
 export const Certs = model('certs', {
   id: t.pk(),
